% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{add_R_core}
\alias{add_R_core}
\title{Add R version to package inventory}
\usage{
add_R_core(pkgs)
}
\arguments{
\item{pkgs}{data.frame of installed packages with columns Package and Version.}
}
\value{
The same data.frame with the R version listed as "R" and the
version in major.minor format (e.g. R 3.2.2).
}
\description{
Add R version to package inventory
}
\keyword{internal}
