% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{check_R_core}
\alias{check_R_core}
\title{Check listed R version against installed}
\usage{
check_R_core(pkgs, R)
}
\arguments{
\item{pkgs}{data.frame read from the csv file.}

\item{R}{If TRUE it will error when the R versions mismatch.
Otherwise it will just give a warning.}
}
\value{
The same data.frame with the R package removed.
}
\description{
Check listed R version against installed
}
\keyword{internal}
