% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{cran_file}
\alias{cran_file}
\title{Get a list of candidate URLs for a certain version of a package}
\usage{
cran_file(package, version, type = get_pkg_type(),
  r_minor = r_minor_version())
}
\arguments{
\item{package}{Name of the package, e.g. \code{jsonlite}.}

\item{version}{Version number of the package, as a string, e.g.
\code{1.0.0}.}

\item{type}{Package type, e.g. \code{binary}, \code{source}, etc.
See the \code{type} argument of \code{utils::install.packages}.}

\item{r_minor}{The minor R version to search for packages for.
Defaults to the currently running R version.}
}
\value{
Character vector or URLs.
}
\description{
Get a list of candidate URLs for a certain version of a package
}
\keyword{internal}

