% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_frame}
\alias{data_frame}
\title{Create a data frame, more robust than \code{data.frame}}
\usage{
data_frame(...)
}
\arguments{
\item{...}{Data frame columns.}
}
\value{
The constructed data frame.
}
\description{
It does not create factor columns.
It recycles columns to match the longest column.
}
\keyword{internal}

