% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\docType{data}
\name{default_cran_mirror}
\alias{default_cran_mirror}
\title{CRAN mirror to use}
\format{An object of class \code{character} of length 1.}
\usage{
default_cran_mirror
}
\description{
The RStudio mirror is in the Amazon cloud, so most times it has
the best response times, and download speed.
}
\keyword{internal}
