% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\docType{data}
\name{dep_types}
\alias{dep_types}
\alias{hard_dep_types}
\title{Dependency types in R DESCRIPTION files}
\format{An object of class \code{character} of length 5.}
\usage{
dep_types

hard_dep_types
}
\description{
Dependency types in R DESCRIPTION files

A dependency is hard if the depended package is required
for installing and/or loading the package.
}
\keyword{internal}

