% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dir_exists}
\alias{dir_exists}
\title{Check if a directory exists}
\usage{
dir_exists(dir)
}
\arguments{
\item{dir}{Directory to check.}
}
\value{
Logical scalar.
}
\description{
Check if a directory exists
}
\keyword{internal}

