% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_urls}
\alias{download_urls}
\title{Get download urls for a bunch of packages}
\usage{
download_urls(pkgs)
}
\arguments{
\item{pkgs}{Character vector of packages, including package
versions, separated by a dash.}
}
\value{
A list of character vectors, a set of URLs for each package.
}
\description{
Get download urls for a bunch of packages
}
\keyword{internal}

