% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{drop_internal}
\alias{drop_internal}
\title{Drop base and recommended packages, and \sQuote{R} from a list
of R packages}
\usage{
drop_internal(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\value{
Character vector of filtered package names.
}
\description{
\sQuote{R} can be included in the DESCRIPTION file, as a dependency,
but we ignore this right now. We also ignore base and recommended
packages, these are supposed to be installed on the system, together
with R.
}
\keyword{internal}
