% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{drop_missing_deps}
\alias{drop_missing_deps}
\title{Drop dependencies that were not included in the snapshot}
\usage{
drop_missing_deps(deps)
}
\arguments{
\item{deps}{A named list of character vectors.}
}
\value{
A (filtered) named list of character vectors.
}
\description{
These are probably not needed for the installed functions
to work.
}
\keyword{internal}

