% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{drop_unneeded_pkgs}
\alias{drop_unneeded_pkgs}
\title{Drop packages that do not need an update}
\usage{
drop_unneeded_pkgs(pkgs)
}
\arguments{
\item{pkgs}{A data frame with columns Package and Version}
}
\value{
A (filtered) data frame with columns Package and Version
}
\description{
If any of the following are true, then a package is needed; otherwise it
unneeded. The package is not installed, or the currently installed version of
the package is not the same as the target version, or the currently installed
package was not built with the current version of R.
}
\keyword{internal}
