% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{filename_from_url}
\alias{filename_from_url}
\title{Extract a file name from a package download URL}
\usage{
filename_from_url(url, pkg)
}
\arguments{
\item{url}{The URL, a character scalar.}

\item{pkg}{The name of the package the URL belongs to.}
}
\value{
Character scalar, the file name.
}
\description{
This is usually just the part after the last slash,
but for https://github.com/cran/* URLs it is a bit trickier.
}
\keyword{internal}
