% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_deps}
\alias{get_deps}
\title{Extract (hard) package dependencies from an R package tarball}
\usage{
get_deps(package_file)
}
\arguments{
\item{package_file}{Path and name of the tarball.}
}
\value{
A character vector of depended packages. Version numbers
  are not included, as we don't need them for the current purposes
  of this package.
}
\description{
Hard dependencies include \code{Imports}, \code{Depends} and
\code{LinkingTo}.
}
\keyword{internal}

