% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{get_description}
\alias{get_description}
\title{Extract and read the DESCRIPTION file from an R package tarball}
\usage{
get_description(package_file)
}
\arguments{
\item{package_file}{Path and name of the tarball.}
}
\value{
A named list of DESCRIPTION fields.
}
\description{
Extract and read the DESCRIPTION file from an R package tarball
}
\keyword{internal}
