% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{get_package_sources}
\alias{get_package_sources}
\title{Check where a package was installed from}
\usage{
get_package_sources(pkgs)
}
\arguments{
\item{pkgs}{The matrix of parsed DESCRIPTION files,
out of `installed.packages`.}
}
\value{
Two columns, source and link
}
\description{
Check where a package was installed from
}
\keyword{internal}
