% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%||\%}
\alias{\%||\%}
\title{LHS if not \code{NULL}, otherwise RHS}
\usage{
l \%||\% r
}
\arguments{
\item{l}{LHS.}

\item{r}{RHS.}
}
\value{
LHS if not \code{NULL}, otherwise RHS.
}
\description{
LHS if not \code{NULL}, otherwise RHS
}
\keyword{internal}

