% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{install_order}
\alias{install_order}
\title{Topological order of the packages}
\usage{
install_order(graph)
}
\arguments{
\item{graph}{A named list of character vectors, interpreted as
an adjacnecy list. If \code{A->B} then package \code{A} depends
on package \code{B}, so package \code{B} must be loaded before
package \code{A}.}
}
\value{
Character vector of package names in an order that
  can be used to install them.
}
\description{
This is the correct installation order.
}
\keyword{internal}

