% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deps.R
\name{parse_deps}
\alias{parse_deps}
\title{Parse a DESCRIPTION dependency field}
\usage{
parse_deps(type, deps)
}
\arguments{
\item{type}{Field name, e.g. \code{Imports}.}

\item{deps}{The value of the field.}
}
\value{
A data frame with three columns: \code{type},
  \code{package} and \code{version}.
}
\description{
Parse a DESCRIPTION dependency field
}
\keyword{internal}

