% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_from_filename}
\alias{pkg_from_filename}
\title{Extract the package name from a package tarball path or filename}
\usage{
pkg_from_filename(path)
}
\arguments{
\item{path}{The package tarball path(s).}
}
\value{
Package name(s).
}
\description{
Extract the package name from a package tarball path or filename
}
\keyword{internal}
