% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/urls.R
\name{r_minor_version}
\alias{r_minor_version}
\title{Extract the minor version of the running R}
\usage{
r_minor_version()
}
\description{
This is needed to calculate possible R package locations
for downloads.
}
\keyword{internal}

