% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore.R
\name{restore}
\alias{restore}
\title{Restore (=install) certain CRAN package versions}
\usage{
restore(from = "packages.csv", R = TRUE, ...)
}
\arguments{
\item{from}{Name of a file created by \code{\link{snap}}.
Alternatively a data frame with columns \code{Package} and
\code{Version}.}

\item{R}{If TRUE the target version of R must match.
Otherwise it will only give a warning.}

\item{...}{Additional arguments, passed to \code{install.packages}.}
}
\description{
Functions that were not installed from CRAN will not be restored,
they will be ignored with a warning. The pkgsnap package itself is
also ignored as it must be installed to run this function.
}

