% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshot.R
\name{snap}
\alias{snap}
\title{Write installed package versions to a file}
\usage{
snap(to = "packages.csv", lib.loc = NULL, recommended = FALSE)
}
\arguments{
\item{to}{File to write the package versions to,
defaults to \code{packages.csv}. If it is NULL,
to output file is created and the result is returned
as a data frame.}

\item{lib.loc}{character vector describing the 
location of R library trees to search through, or 
\code{NULL} for all known trees (see 
\code{\link[base]{.libPaths}}).}

\item{recommended}{if TRUE then recommended packages
will be included in the snapshot.}
}
\value{
A two columns data frame, invisibly if it was
  written to a file.
}
\description{
Base and recommended packages are omitted.
Packages that were installed from non-CRAN sources are
also included, but you won't be able to restore them
with \code{\link{restore}}.
}
\details{
The output file will have two columns, package name
and package version.
}
\examples{
snap(to = tmp <- tempfile())

head(read.csv(tmp))
}

