% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_pkg_names_versions}
\alias{split_pkg_names_versions}
\title{Split pkg-version into data frame with pkg and version}
\usage{
split_pkg_names_versions(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package versions in the
\code{packagename-version} format.}
}
\value{
Data frame with columns \code{name} and \code{version}.
}
\description{
Split pkg-version into data frame with pkg and version
}
\keyword{internal}

