% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{str_trim}
\alias{str_trim}
\title{Trim whitespace from the beginning and end of a string}
\usage{
str_trim(x)
}
\arguments{
\item{x}{Input string or character vector.}
}
\value{
Trimmed character vector.
}
\description{
Trim whitespace from the beginning and end of a string
}
