% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{try_download}
\alias{try_download}
\title{Try to download a file}
\usage{
try_download(url, dest_file)
}
\arguments{
\item{url}{Download URL.}

\item{dest_file}{Where to put the downloaded file.}
}
\value{
\code{TRUE} if the download was successful, \code{FALSE}
  otherwise.
}
\description{
Try to download a file
}
\keyword{internal}
